requirements.txt
fastapi
uvicorn
pandas
joblib
python-dotenv
requests
pyarrow
